Phase 4 non exam recurring delete mod (#338)
* Phase 4 - Booking Delete History

For reporting purposes, booking recurring events will not be deleted after the current date of when the delete function is being called.

This PR also contains a fix for a bug found where if the edit series toggle is turned on, the rescheduling button is hidden along with the date and time incrementer.

* Phase 4 - Sonar Qube: Exam Inventory Table

Changed initialization of length_of_invigilator array from 0 to null.

* Phase 4 - Sonar Qube:Appointments Store Module

Removed un-used state variable from sendToQueue function.

* Phase 4 - Sonar Qube: Booking Blackout Modal

removed all_local_rooms from modal set function.

+++ api/app/resources/bookings/booking/booking_recurring_delete.py
@@ -18,25 +18,37 @@ from app.models.bookings import Booking
 from app.schemas.bookings import BookingSchema
 from app.models.theq import CSR
 from qsystem import api, db, oidc
+from datetime import datetime, timedelta, date
+import pytz
 
 
 @api.route("/bookings/recurring/<string:id>", methods=["DELETE"])
 class BookingRecurringDelete(Resource):
 
     booking_schema = BookingSchema
+    timezone = pytz.timezone("US/Pacific")
 
     @oidc.accept_token(require_token=True)
     def delete(self, id):
+
+        today = datetime.today()
+        string_today = today.strftime('%Y-%m-%d')
 
         print("==> In the python DELETE /bookings/recurring/<id> endpoint")
 
         csr = CSR.find_by_username(g.oidc_token_info['username'])
 
-        bookings = Booking.query.filter_by(recurring_uuid=id).all()
+        bookings = Booking.query.filter_by(recurring_uuid=id)\
+                                .filter(db.func.date(Booking.start_time) >= string_today)\
+                                .all()
 
         for booking in bookings:
             if booking.office_id != csr.office_id and csr.liaison_designate != 1:
                 abort(404)
+
+            if booking.start_time.year == today.year and booking.start_time.month == today.month \
+                    and booking.start_time.day == today.day and booking.start_time.hour <= 5:
+                continue
 
             db.session.delete(booking)
             db.session.commit()
